EDM-5702: Use per-stream output limits in hook executor
Replace the shared-budget output writer with per-stream limits
(max/2 bytes each) and fold runCmd into executer.go.
